backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication
bcrypt==4.1.1
python-jose[cryptography]==3.3.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Vector database
qdrant-client==1.7.0

# AI/LLM
openai==1.6.1

# Utilities
httpx==0.25.2
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Rate limiting
slowapi==0.1.9
